RE #4632 Deleting inexistant CustomTable record should return an error

-1: Ro Bot

Merged-on: https://opencell.assembla.com/code/meveo/git/merge_requests/8126859

+++ opencell-api/src/main/java/org/meveo/api/custom/CustomTableApi.java
@@ -183,6 +183,12 @@ public class CustomTableApi extends BaseApi {
         if (dto.getValues() == null || dto.getValues().isEmpty()) {
             customTableService.remove(cet.getDbTablename());
         } else {
+            Map<Boolean, List<CustomTableRecordDto>> partitionedById = dto.getValues().stream().collect(Collectors.partitioningBy(x->x.getValues().get(FIELD_ID)!=null));
+            List<CustomTableRecordDto> valuesWithoutIds = partitionedById.get(false);
+            
+            if (!valuesWithoutIds.isEmpty()) {
+                throw new ValidationException(valuesWithoutIds.size() + " record(s) to remove are missing the IDs.");
+            }
             Set<Long> ids = extractIds(dto);
             customTableService.remove(cet.getDbTablename(), ids);
         }
